Streamlined Moderation steps, changed "user" to "community member"
User was a carry over from adapting Rust's IRC guidelines. It was
not used consistently, and did not lend itself to the scope of our
CoC. Updated all instances of "user" to "community member" to
better reflect our scope.

Also changed wording in second step in Moderation to reflect that
the banning applies to the community member who committed the
violation. This seems obvious, but moving the immediate ban step
to below the reporting step meant that it could have read as the
community member who reported the violation may be banned.

Removed use of "We" from Moderation steps as it was inconsistent
with the rest of the CoC.